creating custon error, groups and players functionalitys

+++ src/storage/group/groupCreate.ts
@@ -1,5 +1,6 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "@storage/storageConfig";
+import { AppError } from "@utils/AppError";
 import { groupsGetAll } from "./groupsGetAll";
 
 export async function groupCreate(newGroup: string) {
@@ -7,11 +8,17 @@ export async function groupCreate(newGroup: string) {
 
         const storedGroups = await groupsGetAll(); /* Pegando todos os grupos ja criados */
 
+        const groupAlreadyExists = storedGroups.includes(newGroup)
+
+        if (groupAlreadyExists) {
+            throw new AppError('Já existe um grupo cadastrado com esse nome.');
+        }
+
         const storage = JSON.stringify([...storedGroups, newGroup]) /* transformando tudo em string */
 
         await AsyncStorage.setItem(GROUP_COLLECTION, storage); /* salvando todos os grupos */
 
     } catch (error) {
-        throw error;
+        throw error
     }
 }
